Fix Wsign-conversion warning

+++ src/megachatapi_impl.cpp
@@ -8908,7 +8908,7 @@ MegaChatMessagePrivate::MegaChatMessagePrivate(const Message &msg, Message::Stat
                     megaHandleList->addMegaHandle(callEndInfo->participants[i]);
                 }
 
-                priv = callEndInfo->duration;
+                priv = static_cast<int>(callEndInfo->duration);
                 mCode = MegaChatMessagePrivate::convertEndCallTermCodeToUI(*callEndInfo);
                 hAction = callEndInfo->callid;
                 delete callEndInfo;
